+++ pkg/cmd/cli/player/selector.go
@@ -1,0 +1,6 @@
+package player
+
+// Selector is panel for selecting music
+type Selector struct {
+	Panel
+}
