--- pkg/cmd/cli/turntable/select_panel.go
@@ -1,6 +1,0 @@
-package turntable
-
-// SelectPanel is panel for selecting music
-type SelectPanel struct {
-	Panel
-}
